+++ xxl-job-admin/src/main/java/com/xxl/job/admin/dao/XxlJobGroupDao.java
@@ -20,9 +20,9 @@ public interface XxlJobGroupDao {
 
     public int update(XxlJobGroup xxlJobGroup);
 
-    public int remove(@Param("id") int id);
+    public int remove(@Param("id") long id);
 
-    public XxlJobGroup load(@Param("id") int id);
+    public XxlJobGroup load(@Param("id") long id);
 
     public List<XxlJobGroup> pageList(@Param("offset") int offset,
                                       @Param("pagesize") int pagesize,
